requirements.txt
astroid==2.3.3
autopep8==1.5
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.10
django-debug-toolbar==2.2
django-filter==2.2.0
django-import-export==2.0.2
djangorestframework==3.11.0
et-xmlfile==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.3
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
PyYAML==5.3
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
typed-ast==1.4.1
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
